+++ NEWS.md
@@ -2,6 +2,7 @@
 - Switch from `merge` to `[[`
 
 # Change to v.0.99.21
+- Replace `order` with `data.table::setorder`
 - Add support for custom identifiers in merge_SE
 
 # Change to v.0.99.20